+++ src/module/argument.ts
@@ -1,8 +1,8 @@
-import Suggestion from './suggestion';
+import ArgumentSuggestion from './argumentSuggestion';
 import { ARGUMENT_TYPES } from './utils/moduleUtils';
 
 export default interface Argument {
   name: string;
   type: ARGUMENT_TYPES;
-  suggestions?: (...params: any) => Suggestion[];
+  suggestions?: (...params: any) => ArgumentSuggestion[];
 }