+++ src/module/argumentSuggestion.ts
@@ -1,4 +1,4 @@
-export default interface Suggestion {
+export default interface ArgumentSuggestion {
   content: string; // what is shown on the suggestion
   icon?: string;
   img?: string;