+++ src/module/commands/_index.ts
@@ -1,20 +1,22 @@
 import Command from '../command';
 import openCompendiumCommand from './openCompendium';
 import runMacroCommand from './runMacro';
-import openSheetByNameCommand from './openSheetByName';
-import openSheetByPlayerCommand from './openSheetByPlayer';
-import showAllowedCommand from './showAllowedCommands';
-import goTabCommand from './goTab';
+import openPcByNameCommand from './openPcByName';
+import openPcByPlayerCommand from './openPcByPlayer';
+import openNpcByNameCommand from './openNpcByName';
+import helpCommand from './help';
+import tabCommand from './goToTab';
 import infoCommand from './info';
 import tokenActiveEffectCommand from './tokenActiveEffect';
 
 const registerCommands = (register: (command: Command, replace: boolean, silentError: boolean) => void) => {
-  register(openSheetByNameCommand, false, true);
-  register(openSheetByPlayerCommand, false, true);
+  register(openPcByNameCommand, false, true);
+  register(openPcByPlayerCommand, false, true);
+  register(openNpcByNameCommand, false, true);
   register(openCompendiumCommand, false, true);
   register(runMacroCommand, false, true);
-  register(showAllowedCommand, false, true);
-  register(goTabCommand, false, true);
+  register(helpCommand, false, true);
+  register(tabCommand, false, true);
   register(infoCommand, false, true);
   register(tokenActiveEffectCommand, false, true);
 };